Add function GTN_EntropyModuleMagicGenerator

+++ src/main/java/com/EvgenWarGold/GregTechNightmare/GregTech/MultiBlock/Generators/EV/Magen/GTN_AirModuleMagicGenerator.java
@@ -109,7 +109,7 @@ public class GTN_AirModuleMagicGenerator extends GTN_MultiBlockBase<GTN_AirModul
     }
 
     @Override
-    public int generate() {
+    public long generate() {
         return generate;
     }
 
@@ -265,7 +265,7 @@ public class GTN_AirModuleMagicGenerator extends GTN_MultiBlockBase<GTN_AirModul
         currentTip.add(
             EnumChatFormatting.GREEN + "Generate: "
                 + EnumChatFormatting.AQUA
-                + formatNumber(tag.getInteger("generate")));
+                + formatNumber(tag.getLong("generate")));
         currentTip.add(
             EnumChatFormatting.GREEN + "Catalyst duration: "
                 + EnumChatFormatting.AQUA
@@ -277,7 +277,7 @@ public class GTN_AirModuleMagicGenerator extends GTN_MultiBlockBase<GTN_AirModul
         int z) {
         super.getWailaNBTData(player, tile, tag, world, x, y, z);
 
-        tag.setInteger("generate", generate);
+        tag.setLong("generate", generate);
 
         if (catalystData != null) {
             tag.setInteger("catalyst", catalystData.getDuration());
