+++ src/main/java/com/EvgenWarGold/GregTechNightmare/GregTech/MultiBlock/Generators/EV/Magen/GTN_EarthModuleMagicGenerator.java
@@ -104,7 +104,7 @@ public class GTN_EarthModuleMagicGenerator extends GTN_MultiBlockBase<GTN_EarthM
     }
 
     @Override
-    public int generate() {
+    public long generate() {
         return generate;
     }
 
@@ -261,7 +261,7 @@ public class GTN_EarthModuleMagicGenerator extends GTN_MultiBlockBase<GTN_EarthM
         currentTip.add(
             EnumChatFormatting.GREEN + "Generate: "
                 + EnumChatFormatting.AQUA
-                + formatNumber(tag.getInteger("generate")));
+                + formatNumber(tag.getLong("generate")));
         currentTip.add(
             EnumChatFormatting.GREEN + "Catalyst duration: "
                 + EnumChatFormatting.AQUA
@@ -273,7 +273,7 @@ public class GTN_EarthModuleMagicGenerator extends GTN_MultiBlockBase<GTN_EarthM
         int z) {
         super.getWailaNBTData(player, tile, tag, world, x, y, z);
 
-        tag.setInteger("generate", generate);
+        tag.setLong("generate", generate);
 
         if (catalystData != null) {
             tag.setInteger("catalyst", catalystData.getDuration());
