+++ src/main/java/com/EvgenWarGold/GregTechNightmare/GregTech/MultiBlock/Generators/EV/Magen/GTN_FireModuleMagicGenerator.java
@@ -108,7 +108,7 @@ public class GTN_FireModuleMagicGenerator extends GTN_MultiBlockBase<GTN_FireMod
     }
 
     @Override
-    public int generate() {
+    public long generate() {
         return generate;
     }
 
@@ -264,7 +264,7 @@ public class GTN_FireModuleMagicGenerator extends GTN_MultiBlockBase<GTN_FireMod
         currentTip.add(
             EnumChatFormatting.GREEN + "Generate: "
                 + EnumChatFormatting.AQUA
-                + formatNumber(tag.getInteger("generate")));
+                + formatNumber(tag.getLong("generate")));
         currentTip.add(
             EnumChatFormatting.GREEN + "Catalyst duration: "
                 + EnumChatFormatting.AQUA
@@ -276,7 +276,7 @@ public class GTN_FireModuleMagicGenerator extends GTN_MultiBlockBase<GTN_FireMod
         int z) {
         super.getWailaNBTData(player, tile, tag, world, x, y, z);
 
-        tag.setInteger("generate", generate);
+        tag.setLong("generate", generate);
 
         if (catalystData != null) {
             tag.setInteger("catalyst", catalystData.getDuration());
